split utils into smaller modules
