+++ packages/next/src/client/index.tsx
@@ -1,18 +1,20 @@
 /* global location */
 import '../build/polyfills/polyfill-module'
+
 import type Router from '../shared/lib/router/router'
+import type {
+  AppComponent,
+  AppProps,
+  PrivateRouteInfo,
+} from '../shared/lib/router/router'
+
 import React from 'react'
 // @ts-expect-error upgrade react types to react 18
 import ReactDOM from 'react-dom/client'
 import { HeadManagerContext } from '../shared/lib/head-manager-context'
 import mitt, { MittEmitter } from '../shared/lib/mitt'
 import { RouterContext } from '../shared/lib/router-context'
-import {
-  AppComponent,
-  AppProps,
-  handleSmoothScroll,
-  PrivateRouteInfo,
-} from '../shared/lib/router/router'
+import { handleSmoothScroll } from '../shared/lib/router/utils/handle-smooth-scroll'
 import { isDynamicRoute } from '../shared/lib/router/utils/is-dynamic'
 import {
   urlQueryToSearchParams,
